Add pushManager and pushNotification for server
